more refactor

+++ systems/energy.go
@@ -356,7 +356,7 @@ func (s *EnergySystem) handleNuggetCollection(world *engine.World, entity engine
 
 	// Clear the active nugget reference to trigger respawn
 	// Use CAS to ensure we only clear if this is still the active nugget
-	s.nuggetSystem.ClearActiveNuggetIfMatches(uint64(entity))
+	s.nuggetSystem.ClearActiveNuggetIfMatches(entity)
 
 	// Move cursor right in ECS
 	cursorPos, ok := world.Positions.Get(s.ctx.CursorEntity)
